1. Добавил возможность прикреплять изображение к статье (используется в представлении кратких версий статей) 2. Добавил редактирование статьи из редактирования пункта меню 3. Добавил выбор "Использовать стандартный фон" по умолчанию на true 4. При сохранении текста из CKEditor4 мы сначала его пробросим в редактор, дабы применились фильтры, а затем достанем 5. Добавил диалог подтверждения удаления статьи 6. Поправил сортировку дат в таблицах 7. В реестре пунктов меню по умолчанию выбирается русский язык 8. Авто-фокус на инпут поиска по таблице 9. Если в диалоге открыт дочерний диалог, то Esc закроет его (поправил во всех местах) 10. Изменил "Внутренний номер группы" на "Ссылка на расписание" и "Ссылка на учебный план"

+++ src/app/components/dialog/article-edtiting-dialog/article-editing-dialog.component.ts
@@ -1,6 +1,6 @@
-import {Component, HostListener, Inject, OnInit} from '@angular/core';
+import {Component, HostListener, Inject, OnInit, ViewChild} from '@angular/core';
 import {AbstractControl, FormControl, FormGroup, Validators} from '@angular/forms';
-import {MAT_DIALOG_DATA, MatDialogRef} from '@angular/material/dialog';
+import {MAT_DIALOG_DATA, MatDialog, MatDialogRef} from '@angular/material/dialog';
 import {ErrorService} from '../../../services/error/error.service';
 import {MatSnackBar} from '@angular/material/snack-bar';
 import {ImageService} from '../../../services/image/image.service';
@@ -11,8 +11,7 @@ import {AppConfig} from '../../../app.config';
 import {TdLoadingService} from '@covalent/core/loading';
 import {transformResultTextToHtml} from "../../common/ckeditor/utils-function";
 import {MatStepper} from "@angular/material/stepper";
-import {DialogService} from "../../../services/dialog/dialog.service";
-import {TokenStorageService} from "../../../services/token/token-storage.service";
+import {ConfirmDialogComponent} from "../../common/dialog/confirm-dialog/confirm-dialog.component";
 
 export interface ArticleEditingDialogData {
   articleId: string;
@@ -25,6 +24,8 @@ export interface ArticleEditingDialogData {
   templateUrl: './article-editing-dialog.component.html'
 })
 export class ArticleEditingDialogComponent implements OnInit {
+
+  @ViewChild("myEditor", { static: false }) myEditor: any;
 
   backgroundImageStyle = 'body { background-image: ' +
     'url(https://internationals.tpu.ru:8080/api/media/img/7CDAEA5F-BC6D-461B-B214-12A975E88A55) }';
@@ -41,6 +42,8 @@ export class ArticleEditingDialogComponent implements OnInit {
 
   currentArticleId: string;
 
+  imageId: string;
+
   loaderName = 'loader';
 
   constructor(
@@ -51,7 +54,8 @@ export class ArticleEditingDialogComponent implements OnInit {
     private imageService: ImageService,
     private articleService: ArticleService,
     private appConfig: AppConfig,
-    private loadingService: TdLoadingService
+    private loadingService: TdLoadingService,
+    private matDialog: MatDialog
   ) {
     this.loadingService.register(this.loaderName);
     this.currentMode = data.mode;
@@ -61,8 +65,13 @@ export class ArticleEditingDialogComponent implements OnInit {
       this.articleService.getArticleById(data.articleId).subscribe(it => {
         if(it) {
           this.generalInfoForm.patchValue(it);
+          if (it.imageId) {
+            this.imageId = it.imageId;
+          }
           if(this.hasBackgroundStyle(it.text)) {
             this.generalInfoForm.patchValue({useBackground: true});
+          } else {
+            this.generalInfoForm.patchValue({useBackground: false});
           }
           this.textControl.patchValue(it.text)
           this.loadingService.resolve(this.loaderName);
@@ -79,7 +88,8 @@ export class ArticleEditingDialogComponent implements OnInit {
     this.generalInfoForm.addControl('topic', new FormControl('', Validators.required));
     this.generalInfoForm.addControl('briefText', new FormControl('', null));
     this.generalInfoForm.addControl('language', new FormControl('', Validators.required));
-    this.generalInfoForm.addControl('useBackground', new FormControl(false, Validators.required));
+    this.generalInfoForm.addControl('image', new FormControl(null, null));
+    this.generalInfoForm.addControl('useBackground', new FormControl(true, Validators.required));
     this.textControl = new FormControl(null, Validators.required);
   }
 
@@ -106,28 +116,32 @@ export class ArticleEditingDialogComponent implements OnInit {
   }
 
   accept(): void {
-    let htmlText = transformResultTextToHtml(this.textControl.value);
-    const useBackground = this.generalInfoForm.get('useBackground').value;
-    if(useBackground && !this.hasBackgroundStyle(htmlText)) {
-      htmlText = this.addBackgroundStyle(htmlText);
-    }
-    const articleInfo: Article = {
-      name: this.get('name').value,
-      topic: this.get('topic').value,
-      briefText: this.get('briefText').value,
-      text: htmlText,
-      language: this.get('language').value
-    };
-    if(this.currentMode === this.mode.CREATE) {
-      this.articleService.create(articleInfo).subscribe(() => {
-        this.dialogRef.close(true);
-      }, error => this.errorService.handleServiceError(error));
-    } else {
-      this.articleService.update(articleInfo, this.currentArticleId).subscribe(
-        () => this.dialogRef.close(true),
-        error => this.errorService.handleServiceError(error)
-      );
-    }
+    // @ts-ignore
+    this.myEditor.instance.setData(this.textControl.value, () => {
+      let htmlText = transformResultTextToHtml(this.textControl.value);
+      const useBackground = this.generalInfoForm.get('useBackground').value;
+      if(useBackground && !this.hasBackgroundStyle(htmlText)) {
+        htmlText = this.addBackgroundStyle(htmlText);
+      }
+      const articleInfo: Article = {
+        name: this.get('name').value,
+        topic: this.get('topic').value,
+        briefText: this.get('briefText').value,
+        text: htmlText,
+        language: this.get('language').value,
+        imageId: this.imageId
+      };
+      if(this.currentMode === this.mode.CREATE) {
+        this.articleService.create(articleInfo).subscribe((articleId) => {
+          this.dialogRef.close(articleId);
+        }, error => this.errorService.handleServiceError(error));
+      } else {
+        this.articleService.update(articleInfo, this.currentArticleId).subscribe(
+          () => this.dialogRef.close(this.currentArticleId),
+          error => this.errorService.handleServiceError(error)
+        );
+      }
+    });
   }
 
   private addBackgroundStyle(text: string): string {
@@ -144,10 +158,23 @@ export class ArticleEditingDialogComponent implements OnInit {
   }
 
   delete() {
-    this.articleService.delete(this.currentArticleId).subscribe(() => {
-      this.snackBar.open('Статья удалена', 'Закрыть', {duration: 3000});
-      this.dialogRef.close(true);
-    }, error => this.errorService.handleServiceError(error))
+    this.matDialog.open(ConfirmDialogComponent, {
+      data: {
+        title: 'Удаление статьи',
+        message: 'Вы уверены, что хотите удалить статью?',
+        acceptButton: 'Удалить',
+        cancelButton: 'Отмена'
+      },
+      width: '500px',
+      disableClose: true
+    }).afterClosed().subscribe(it => {
+      if(it) {
+        this.articleService.delete(this.currentArticleId).subscribe(() => {
+          this.snackBar.open('Статья удалена', 'Закрыть', {duration: 3000});
+          this.dialogRef.close(true);
+        }, error => this.errorService.handleServiceError(error))
+      }
+    })
   }
 
   nextStep(stepper: MatStepper) {
@@ -156,6 +183,30 @@ export class ArticleEditingDialogComponent implements OnInit {
 
   prevStep(stepper: MatStepper) {
     stepper.previous();
+  }
+
+  selectImage(): void {
+    this.loadingService.register(this.loaderName);
+    const file = this.get('image').value;
+    if (/^image\//.test(file.type)) {
+      this.imageService.upload(file).subscribe(it => {
+        if (it) {
+          this.imageId = it;
+          this.loadingService.resolve(this.loaderName);
+        }
+      });
+    } else {
+      this.snackBar.open('Можно загружать только изображения',
+        'Закрыть', {duration: 3000});
+      this.loadingService.resolve(this.loaderName);
+    }
+  }
+
+  openImage(): void {
+    let image = new Image();
+    image.src = this.appConfig.webServiceFullUrl + '/media/img/' + this.imageId;
+    let win = window.open('');
+    win.document.write(image.outerHTML);
   }
 
   @HostListener('window:keyup.esc')
